1.购物车商品数量发生改变时对相关数据进行修改，并对页面进行刷新2.购物车数量超过库存时对用户行为进行提醒，并对页面进行刷新

+++ beststore/src/main/java/com/bs/beststore/web/action/GoodsAction.java
@@ -117,7 +117,7 @@ public class GoodsAction {
 		// 修改购物车数量信息
 		if (session.getAttribute("loginHuman") != null) {
 			Human human = (Human) session.getAttribute("loginHuman");
-			int cartCount = (int) cartBiz.countByHid(human.getHid());
+			Long cartCount = cartBiz.countByHid(human.getHid());
 			session.setAttribute("cartCount", cartCount);
 		}
 		if(color==null || color.equals("")) {
